Refactor email handling and templates
Separate email sending logic and templates into dedicated modules.

* **Add `server/lib/email.js`**: Create a new module for handling email sending using `nodemailer`. Define and export the `sendEmail` function.
* **Add `server/lib/emailTemplates.js`**: Create a new module for email templates. Define and export `registrationTemplate` and `passwordResetTemplate` functions.
* **Modify `server/router_account.js`**: Import `sendEmail` and email templates. Replace existing email sending logic with calls to `sendEmail` and use the new templates. Remove `nodemailer` configuration.
* **Modify `server/router_user.js`**: Import `sendEmail` and email templates. Remove `nodemailer` configuration.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/ZeroCatDev/ZeroCat?shareId=XXXX-XXXX-XXXX-XXXX).
