+++ server/lib/email.js
@@ -1,0 +1,25 @@
+const nodemailer = require('nodemailer');
+
+const transporter = nodemailer.createTransport({
+  service: process.env.MAIL_SERVICE,
+  secure: true,
+  auth: {
+    user: process.env.MAIL_USER,
+    pass: process.env.MAIL_PASS,
+  },
+});
+
+const sendEmail = async (to, subject, html) => {
+  try {
+    await transporter.sendMail({
+      from: `${process.env.SITE_NAME} <${process.env.MAIL_FROM}>`,
+      to,
+      subject,
+      html,
+    });
+  } catch (error) {
+    console.error('Error sending email:', error);
+  }
+};
+
+module.exports = { sendEmail };
